% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addnum.R
\name{addnum}
\alias{addnum}
\title{Compute (a+b)^m}
\usage{
addnum(x, y, m)
}
\arguments{
\item{m}{A numeric value representing the exponent.}

\item{a}{A numeric value representing the first addend.\code{\link[base]{sum}}}

\item{b}{A numeric value representing the second addend.}
}
\value{
A numeric value that is the result of (a+b)^m.
}
\description{
This function calculates the m-th power of the sum of a and b, i.e., (a+b)^m.
}
\examples{
# Calculate (2+3)^2, which should return 25
result <- my_sum(2, 3, 2)
print(result)

# Calculate (-1+4)^3, which should return 125
print(my_sum(-1, 4, 3))
}
