% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFTexp.R
\name{AFTexp}
\alias{AFTexp}
\title{create an exponential AFT model}
\usage{
AFTexp(time, delta, x, data)
}
\arguments{
\item{time}{a numeric vector representing time}

\item{delta}{a numeric vector representing survival status}

\item{x}{a vector representing the predictor}

\item{data}{the dataset the vectors reside in}
}
\value{
exponential AFT model
}
\description{
creates an exponential AFT model with 1 predictor in a simple function
}
\examples{
library(survival)
data(rats)
mod1 <- AFTexp(rats$time,rats$status,rats$rx,rats)
mod1
}
\author{
Vlad
}
