% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFTlog.R
\name{AFTlog}
\alias{AFTlog}
\title{create a log-logistic AFT model}
\usage{
AFTlog(time, delta, x, data)
}
\arguments{
\item{time}{a numeric vector representing time}

\item{delta}{a numeric vector representing survival status}

\item{x}{a vector representing the predictor}

\item{data}{the dataset the vectors reside in}
}
\value{
log-logistic AFT model
}
\description{
creates a log logistic AFT model with 1 predictor in a simple function
}
\examples{
library(survival)
data(rats)
mod1 <- AFTlog(rats$time,rats$status,rats$rx,rats)
mod1
}
\author{
Vlad
}
