% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFTweib.R
\name{AFTweib}
\alias{AFTweib}
\title{create a weibull AFT model}
\usage{
AFTweib(time, delta, x, data)
}
\arguments{
\item{time}{a numeric vector representing time}

\item{delta}{a numeric vector representing survival status}

\item{x}{a vector representing the predictor}

\item{data}{the dataset the vectors reside in}
}
\value{
weibull AFT model
}
\description{
creates a weibull AFT model with 1 predictor in a simple function
}
\examples{
library(survival)
data(rats)
mod1 <- AFTweib(rats$time,rats$status,rats$rx,rats)
mod1
}
\author{
Vlad
}
